Change - primaryKey zu int
* Alle Variablen und Rückgaben von Long zu Integre geändert, da in der
Datenbank dieser auch als Integre hinterlegt ist.

+++ src/main/java/de/edlly/db/SQLite.java
@@ -4,17 +4,17 @@ import java.sql.Connection;
 
 public abstract class SQLite {
     private SQLiteConnect sqlConnection;
-    private long primaryKey;
+    private int primaryKey;
 
     public SQLite() {
 
     }
     
-    public long getPrimaryKey() {
+    public int getPrimaryKey() {
         return primaryKey;
     }
 
-    public void setPrimaryKey(long id) {
+    public void setPrimaryKey(int id) {
         this.primaryKey = id;
     }
